+++ html/search/groups_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['basic_20declarations_777',['Basic declarations',['../group___basic.html',1,'']]]
+  ['basic_20declarations_780',['Basic declarations',['../group___basic.html',1,'']]]
 ];