+++ html/search/groups_5.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['operator_20expressions_791',['Operator expressions',['../group___operators.html',1,'']]]
+  ['operator_20expressions_794',['Operator expressions',['../group___operators.html',1,'']]]
 ];