+++ html/search/related_1.js
@@ -1,4 +1,5 @@
 var searchData=
 [
-  ['hamiltonian_766',['Hamiltonian',['../class_pomerol_1_1_hamiltonian_part.html#a96f8faa23a547dfab8df3835dcdea91c',1,'Pomerol::HamiltonianPart']]]
+  ['gfcontainer_766',['GFContainer',['../class_pomerol_1_1_greens_function.html#afb275e0d817fc0840f56bd151d29c9da',1,'Pomerol::GreensFunction::GFContainer()'],['../class_pomerol_1_1_greens_function_part.html#afb275e0d817fc0840f56bd151d29c9da',1,'Pomerol::GreensFunctionPart::GFContainer()']]],
+  ['greensfunction_767',['GreensFunction',['../class_pomerol_1_1_greens_function_part.html#ace32eac3b8562de54b5fde16d4d1df3f',1,'Pomerol::GreensFunctionPart']]]
 ];