+++ html/search/related_2.js
@@ -1,6 +1,4 @@
 var searchData=
 [
-  ['indexcontainer2_3c_20greensfunction_2c_20gfcontainer_20_3e_767',['IndexContainer2&lt; GreensFunction, GFContainer &gt;',['../class_pomerol_1_1_g_f_container.html#a3018e0ac524839bde226d096f4069fae',1,'Pomerol::GFContainer']]],
-  ['indexcontainer4_3c_20threepointsusceptibility_2c_20threepointsusceptibilitycontainer_20_3e_768',['IndexContainer4&lt; ThreePointSusceptibility, ThreePointSusceptibilityContainer &gt;',['../class_pomerol_1_1_three_point_susceptibility_container.html#a1fbfcf5861d095d79180064a60214153',1,'Pomerol::ThreePointSusceptibilityContainer']]],
-  ['indexcontainer4_3c_20twoparticlegf_2c_20twoparticlegfcontainer_20_3e_769',['IndexContainer4&lt; TwoParticleGF, TwoParticleGFContainer &gt;',['../class_pomerol_1_1_two_particle_g_f_container.html#abe4b67093c1f8c2f7c0e07ce5a64870b',1,'Pomerol::TwoParticleGFContainer']]]
+  ['hamiltonian_768',['Hamiltonian',['../class_pomerol_1_1_hamiltonian_part.html#a96f8faa23a547dfab8df3835dcdea91c',1,'Pomerol::HamiltonianPart']]]
 ];