+++ html/search/related_3.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['matsubaracontainer4_3c_20vertex4_20_3e_770',['MatsubaraContainer4&lt; Vertex4 &gt;',['../class_pomerol_1_1_vertex4.html#a0d75b1839b6e05285ad861cbfd8fbd21',1,'Pomerol::Vertex4']]]
+  ['indexcontainer2_3c_20greensfunction_2c_20gfcontainer_20_3e_769',['IndexContainer2&lt; GreensFunction, GFContainer &gt;',['../class_pomerol_1_1_g_f_container.html#a3018e0ac524839bde226d096f4069fae',1,'Pomerol::GFContainer']]],
+  ['indexcontainer4_3c_20threepointsusceptibility_2c_20threepointsusceptibilitycontainer_20_3e_770',['IndexContainer4&lt; ThreePointSusceptibility, ThreePointSusceptibilityContainer &gt;',['../class_pomerol_1_1_three_point_susceptibility_container.html#a1fbfcf5861d095d79180064a60214153',1,'Pomerol::ThreePointSusceptibilityContainer']]],
+  ['indexcontainer4_3c_20twoparticlegf_2c_20twoparticlegfcontainer_20_3e_771',['IndexContainer4&lt; TwoParticleGF, TwoParticleGFContainer &gt;',['../class_pomerol_1_1_two_particle_g_f_container.html#abe4b67093c1f8c2f7c0e07ce5a64870b',1,'Pomerol::TwoParticleGFContainer']]]
 ];